archivematicaCommon: Move get_file_checksum to archivematicaFunctions
Remove checksummingTools from externals, since all it contains is a single
function. Also, remove unused archivematicaMCPFileUUID
